backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas>=2.2.0
scikit-learn>=1.3.2
joblib>=1.3.2
SQLAlchemy>=2.0.23
pymysql>=1.1.0
cryptography>=41.0.7
python-dotenv>=1.0.0
bcrypt>=4.0.1

